now 50% less unsound! don't rely on set_ptr_value in slimbox_unsize!
Sadly, we can't infer the type now, so the one-argument form of the
macro is now gone. `Box<T>` to `SlimBox<T>` still needs set_ptr_value
so that's now gated behind either the `nightly` feature (which requires
nightly) or the `unsound_stable` feature. If both are enabled we still
use the not-actually-unsound version from nightly.